requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Jinja2>=3.1.2

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.10.0.84
pymupdf>=1.24.0
pytesseract>=0.3.10

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.65.0
aiofiles>=23.0.0